pyproject.toml
[project]
name = "intelligent-warehouse-multi-agent-web-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "isort>=8.0.0",
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]